infrastructure/src/main/resources/logger/requirements.txt
google-cloud-bigquery==2.20.0
tensorflow==2.5.0
pillow==9.0.0
tf-agents==0.8.0